HybridPart Improvements
Switched from 5d to 6d coordinate systems (5d was non-deterministic
when primary z component was 0 -- there does exist a 5d version,
but this is more work to implement)

Added canonicalization for planes, cylinders, and lines to remove
bad outliers.

Overfitting to frame guide is working except for the mask, which
still has difficulties on the side bars.