KNN resampling experiments.
Experiments with using a KNN to resample the image domain for better
lerning.

Started by using KNN interpolation to get an idea of how well we can
expect to predict masks using the number of samples we have.
Experimented with changing the number of neighbors averaged, the number
of reference samples, and the number of samples.

Found that more than 1 neighbor (either straight or weighted average)
produces jagged boundaries. 1000 - 1500 samples appears sufficient for
the frame guide.

However, I also found that if we grid-sample the KNN and then blur it
with a guassian filter, we can clean up the edges and get away with
slightly fewer samples (at the cost of rounded corners). Using sigma=3
and 750 samples on a 200x200 grid appears to work well, though trying to
reproduce this outside of a grid search didn't work for some reason.

Both KNN and Guassian Blur are implemented by pytorch geometric and
torchvision, respectively, on the GPU, so this resampling is something
that could be done entirely on the GPU at training time.

I also found sample_grid and interpolate functions in pytorch which
could be useful.

In particular, trying to allocate the entire 200x200 grid for learning
is too much for my laptop GPU, but resampling it every epoch could be
plausible (either downscaling or taking random sub-samples).